Reorganized data and python into one folder
