+++ Explore-data-web-page/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <title>The Local Elections Database Project</title>
-    <link rel="stylesheet" href="../style.css">
+    <link rel="stylesheet" href="styles/style.css">
     <link rel="stylesheet" href="https://use.typekit.net/qnf2gwf.css">
     <script src="maptoggle.js" defer></script>
     <script src="https://d3js.org/d3.v5.min.js"></script>
