+++ citation-page/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <title>The Local Elections Database Project</title>
-    <link rel="stylesheet" href="../style.css">
+    <link rel="stylesheet" href="styles/style.css">
     <link rel="stylesheet" href="https://use.typekit.net/qnf2gwf.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 </head>
