+++ gender-web-page-county-leg/index.html
@@ -2,13 +2,13 @@
 <html lang="en">
 <head>
     <title>The Local Elections Database Project</title>
-    <link rel="stylesheet" href="../style.css">
+    <link rel="stylesheet" href="styles/style.css">
     <link rel="stylesheet" href="https://use.typekit.net/qnf2gwf.css">
     <script src="maptoggle.js" defer></script>
     <script src="https://d3js.org/d3.v5.min.js"></script>
     <script src="https://d3js.org/topojson.v3.min.js"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../stat_style.css">
+    <link rel="stylesheet" href="styles/stat_style.css">
 </head>
 <body>
     <div class="container">
